+++ tests/entrypoints/serve/middleware/test_authentication_middleware.py
@@ -11,7 +11,7 @@ from starlette.responses import JSONResponse
 from starlette.routing import Route
 from starlette.testclient import TestClient
 
-from vllm.entrypoints.openai.api_server import register_api_routers
+from vllm.entrypoints.launchers.api_server.routers import register_api_routers
 from vllm.entrypoints.serve.middleware.authenticate import (
     GUARDED_PREFIX,
     AuthenticationMiddleware,
@@ -57,113 +57,6 @@ def _create_app_with_mock_routes(routes: list[tuple[str, list[str]]]) -> FastAPI
             include_in_schema=False,
         )
     return app
-
-
-# ---------------------------------------------------------------------------
-# Documented endpoint lists (must be kept in sync with official docs)
-# These lists are used for explicit testing and documentation coverage.
-# ---------------------------------------------------------------------------
-
-DOCUMENTED_PROTECTED_ENDPOINTS: list[tuple[str, list[str]]] = [
-    ("/v1/models", ["GET"]),
-    ("/v1/chat/completions", ["POST"]),
-    ("/v1/chat/completions/batch", ["POST"]),
-    ("/v1/chat/completions/render", ["POST"]),
-    ("/v1/chat/completions/derender", ["POST"]),
-    ("/v1/completions", ["POST"]),
-    ("/v1/completions/render", ["POST"]),
-    ("/v1/completions/derender", ["POST"]),
-    ("/v1/embeddings", ["POST"]),
-    ("/v1/audio/transcriptions", ["POST"]),
-    ("/v1/audio/translations", ["POST"]),
-    ("/v1/messages", ["POST"]),
-    ("/v1/messages/count_tokens", ["POST"]),
-    ("/v1/responses", ["POST"]),
-    ("/v1/responses/{response_id}", ["GET"]),
-    ("/v1/responses/{response_id}/cancel", ["POST"]),
-    ("/v1/score", ["POST"]),
-    ("/v1/rerank", ["POST"]),
-    ("/v1/load_lora_adapter", ["POST"]),
-    ("/v1/unload_lora_adapter", ["POST"]),
-    ("/inference/v1/generate", ["POST"]),
-    ("/v2/embed", ["POST"]),
-    ("/v2/rerank", ["POST"]),
-]
-
-DOCUMENTED_UNPROTECTED_ENDPOINTS: list[tuple[str, list[str]]] = [
-    ("/invocations", ["POST"]),
-    ("/generative_scoring", ["POST"]),
-    ("/pooling", ["POST"]),
-    ("/classify", ["POST"]),
-    ("/score", ["POST"]),
-    ("/rerank", ["POST"]),
-    ("/pause", ["POST"]),
-    ("/resume", ["POST"]),
-    ("/is_paused", ["GET"]),
-    ("/abort_requests", ["POST"]),
-    ("/scale_elastic_ep", ["POST"]),
-    ("/is_scaling_elastic_ep", ["GET"]),
-    ("/init_weight_transfer_engine", ["POST"]),
-    ("/update_weights", ["POST"]),
-    ("/get_world_size", ["GET"]),
-    ("/tokenize", ["POST"]),
-    ("/detokenize", ["POST"]),
-    ("/health", ["GET"]),
-    ("/ping", ["GET"]),
-    ("/version", ["GET"]),
-    ("/load", ["GET"]),
-    ("/tokenizer_info", ["GET"]),
-    ("/server_info", ["GET"]),
-    ("/reset_prefix_cache", ["POST"]),
-    ("/reset_mm_cache", ["POST"]),
-    ("/reset_encoder_cache", ["POST"]),
-    ("/sleep", ["POST"]),
-    ("/wake_up", ["POST"]),
-    ("/is_sleeping", ["GET"]),
-    ("/collective_rpc", ["POST"]),
-    ("/start_profile", ["POST"]),
-    ("/stop_profile", ["POST"]),
-]
-
-# Endpoints that are documented but intentionally not auto-discovered
-# (e.g., they are only registered in development mode or are internal).
-# These are excluded from coverage checks to avoid false positives.
-missing_documented_endpoints = {
-    "/redoc",
-    "/weight_info",
-    "/start_draft_weight_update",
-    "/openapi.json",
-    "/update_weight_version",
-    "/docs/oauth2-redirect",
-    "/docs",
-    "/finish_weight_update",
-    "/start_weight_update",
-    "/metrics",
-}
-
-# Auto-discovered endpoints that are not in the documented lists,
-# but are known to be legitimate (e.g., they are task-specific or legacy).
-# These are excluded from the "extra endpoints" check.
-missing_auto_discovered_endpoints = {
-    "/score",
-    "/v1/load_lora_adapter",
-    "/v2/rerank",
-    "/start_profile",
-    "/rerank",
-    "/v1/score",
-    "/v2/embed",
-    "/stop_profile",
-    "/pooling",
-    "/v1/unload_lora_adapter",
-    "/v1/embeddings",
-    "/classify",
-    "/tokenizer_info",
-    "/v1/rerank",
-}
-
-# ---------------------------------------------------------------------------
-# Fixture: auto-discovered routes per task
-# ---------------------------------------------------------------------------
 
 
 @pytest.fixture(params=get_args(SupportedTask))
@@ -243,112 +136,3 @@ def test_auto_discovered_unprotected_routes_no_auth(task_routes):
         assert resp.status_code == 200, (
             f"[{task}] {test_method} {test_path} should be accessible without token"
         )
-
-
-# ---------------------------------------------------------------------------
-# Tests for documented endpoint lists
-# ---------------------------------------------------------------------------
-
-
-def test_documented_protected_endpoints_require_auth():
-    app = _create_app_with_mock_routes(DOCUMENTED_PROTECTED_ENDPOINTS)
-    for path_template, methods in DOCUMENTED_PROTECTED_ENDPOINTS:
-        client = TestClient(app)
-        test_path = generate_test_path(path_template)
-        test_method = methods[0] if methods else "GET"
-
-        resp = client.request(test_method, test_path)
-        assert resp.status_code == 401, (
-            f"{test_method} {test_path} should reject missing token"
-        )
-
-        resp = client.request(
-            test_method, test_path, headers={"Authorization": "Bearer wrong"}
-        )
-        assert resp.status_code == 401, (
-            f"{test_method} {test_path} should reject invalid token"
-        )
-
-        resp = client.request(
-            test_method, test_path, headers={"Authorization": "Bearer valid-token"}
-        )
-        assert resp.status_code == 200, (
-            f"{test_method} {test_path} should accept valid token"
-        )
-
-
-def test_documented_unprotected_endpoints_no_auth():
-    app = _create_app_with_mock_routes(DOCUMENTED_UNPROTECTED_ENDPOINTS)
-
-    for path_template, methods in DOCUMENTED_UNPROTECTED_ENDPOINTS:
-        client = TestClient(app)
-        test_path = generate_test_path(path_template)
-        test_method = methods[0] if methods else "GET"
-
-        resp = client.request(test_method, test_path)
-        assert resp.status_code == 200, (
-            f"{test_method} {test_path} should be accessible without token"
-        )
-
-
-# ---------------------------------------------------------------------------
-# Coverage comparison: documented vs. auto-discovered (across all tasks)
-# ---------------------------------------------------------------------------
-
-
-def _collect_all_auto_routes(monkeypatch) -> set[str]:
-    """Collect all auto-discovered route paths across all SupportedTasks."""
-    monkeypatch.setenv("VLLM_SERVER_DEV_MODE", "1")
-    all_paths: set[str] = set()
-    for task in get_args(SupportedTask):
-        app = FastAPI()
-        args = Namespace()
-        app.state = Namespace()
-        app.state.args = args
-        register_api_routers(args, app, supported_tasks=(task,))
-        routes = get_all_http_routes(app)
-        all_paths.update({path for path, _ in routes})
-    return all_paths
-
-
-def test_no_missing_documented_endpoints(monkeypatch):
-    """Verify that every documented endpoint is registered by at
-    least one SupportedTask under development mode."""
-    auto_paths = _collect_all_auto_routes(monkeypatch)
-    documented = {
-        path
-        for path, _ in DOCUMENTED_PROTECTED_ENDPOINTS + DOCUMENTED_UNPROTECTED_ENDPOINTS
-    }
-
-    missing = documented - auto_paths - missing_auto_discovered_endpoints
-    assert not missing, (
-        "The following documented endpoints were not registered by any task: "
-        f"{sorted(missing)}"
-    )
-
-
-def test_report_extra_auto_discovered_endpoints(monkeypatch):
-    """
-    Report (and fail if any) auto-discovered endpoints that are not present
-    in the documented lists. This helps keep documentation up-to-date.
-    """
-    auto_paths = _collect_all_auto_routes(monkeypatch)
-    documented = {
-        path
-        for path, _ in DOCUMENTED_PROTECTED_ENDPOINTS + DOCUMENTED_UNPROTECTED_ENDPOINTS
-    }
-
-    extra = auto_paths - documented - missing_documented_endpoints
-    if extra:
-        print("\nExtra auto-discovered endpoints not in documented lists:")
-        for path in sorted(extra):
-            print(f"  {path}")
-
-    # Make the test fail if there are extra endpoints, forcing documentation update.
-    assert not extra, (
-        f"Auto-discovered routes not covered by documented lists. "
-        f"Please update the documentation at "
-        f"'docs/serving/online_serving/README.md' and refer to "
-        f"'https://github.com/vllm-project/vllm/blob/main/docs/usage/security.md#api-key-authentication-limitations' "  # noqa: E501
-        f"to maintain consistency. Extra endpoints: {sorted(extra)}"
-    )
